+++ package.json
@@ -1,11 +1,13 @@
 {
   "name": "outils",
-  "version": "1.3.1",
+  "version": "1.3.2",
   "description": "前端业务代码工具库",
   "main": "min/outils.min.js",
   "scripts": {
     "build": "node script/build.js",
-    "test": "karma start script/karma.conf.js"
+    "build-copy": "node script/build-copy.js",
+    "build-pack": "node script/build-pack",
+    "test": "npm run build-pack & karma start script/karma.conf.js"
   },
   "repository": {
     "type": "git",
@@ -37,7 +39,5 @@
     "rimraf": "^2.6.2",
     "webpack": "^3.8.1"
   },
-  "dependencies": {
-    "recursive-readdir-sync": "^1.0.6"
-  }
+  "dependencies": {}
 }
