+++ src/array/arrayEqual.js
