+++ src/regexp/isPhoneNum.js
@@ -5,7 +5,7 @@
  * @return {Boolean} 
  */
 function isPhoneNum(str) {
-    return /^(0|86|17951)?(13[0-9]|15[012356789]|17[678]|18[0-9]|14[57]|19[0-9])[0-9]{8}$/.test(str)
+    return /^(\+?0?86\-?)?1[3456789]\d{9}$/.test(str)
 }
 
 module.exports = isPhoneNum
