+++ test/object.test.js
@@ -1,5 +1,20 @@
 describe('Object API:', function () {
     describe('#deepClone()', function () {
+        it(`person deepEqual outils.deepClone(null) should return true`, function () {
+            let person = null
+            assert.deepEqual(person, outils.deepClone(person))
+        });
+
+        it(`person deepEqual outils.deepClone('SlaneYang') should return true`, function () {
+            let person = 'SlaneYang'
+            assert.deepEqual(person, outils.deepClone(person))
+        });
+
+        it(`person deepEqual outils.deepClone(new Date()) should return true`, function () {
+            let date = new Date()
+            assert.deepEqual(date, outils.deepClone(date))
+        });
+
         it(`person deepEqual outils.deepClone(person) should return true`, function () {
             let person = {
                 name: "user",
